0.1.10

+++ src/lib/reports/lib/generator/backup.ts
@@ -17,24 +17,28 @@ export async function Backup (full_store_folder: string, full_backup_folder: str
         logger.log(`[Reports] Folder ${chalk.grey(full_project_backup_path)} for project ${chalk.grey(project)} created`, "dev");
     }  
 
-    const files = fs.readdirSync(full_project_allure_path);
+    if (fs.existsSync(full_project_allure_path) === true) {
 
-    for (const file_path of files) {
+        const files = fs.readdirSync(full_project_allure_path);
 
-        const full_file_src_path = path.resolve(full_project_allure_path, file_path);
-        const full_file_dest_path = path.resolve(full_project_backup_path, file_path);
-        const stat = fs.statSync(full_file_src_path);
+        for (const file_path of files) {
 
-        if (stat.isFile() === true) {
-            fs.copyFileSync(full_file_src_path, full_file_dest_path);
-            logger.log(`[Reports] File ${chalk.grey(full_file_src_path)} backup to ${chalk.grey(full_file_dest_path)}`, "dev");
+            const full_file_src_path = path.resolve(full_project_allure_path, file_path);
+            const full_file_dest_path = path.resolve(full_project_backup_path, file_path);
+            const stat = fs.statSync(full_file_src_path);
+
+            if (stat.isFile() === true) {
+                fs.copyFileSync(full_file_src_path, full_file_dest_path);
+                logger.log(`[Reports] File ${chalk.grey(full_file_src_path)} backup to ${chalk.grey(full_file_dest_path)}`, "dev");
+            }
+
         }
 
-    }
+        fs.rmdirSync(full_project_allure_path, {
+            recursive: true
+        });
 
-    fs.rmdirSync(full_project_allure_path, {
-        recursive: true
-    });
+    }
 
     logger.log(`[Reports] Backup test files for project ${chalk.grey(project)} complete`, "dev");
